style: refactor models coming from runner
