+++ backend/api/models/__init__.py
@@ -1,3 +1,6 @@
 from .courseModel import *  # noqa: W0611
 from .exerciseModel import *  # noqa: W0611
 from .sessionModel import *  # noqa: W0611
+from .submissionModel import *  # noqa: W0611
+from .testModel import *  # noqa: W0611
+from .testResultModel import *  # noqa: W0611
